==> 6. Análisis de sentimiento.ipynb <==

###  <span style="color:orange"> ANÁLISIS DE SENTIMIENTOS </span>
___
La última exploración que haremos será una aproximación al análisis de sentimiento en idioma español. Esta técnica consiste en extraer información de la subjetividad presente en una serie de textos o documentos, normalmente centrado en la extracción de emociones, discriminando entre tres valores fundamentales: positivo, negativo y neutral.

La gran distancia de recursos existentes para modelos en lengua inglesa y española nos impide en este apartado del uso de paquetes tan completos como TextBlob o las funciones de análisis de sentimiento de word2vec, dos de las herramientas más comunes y eficientes en el mundo del análisis de sentimiento.

Para nuestro ejercicio utilizaremos el recurso sentiment-analysis-spanish de Hugo J. Bello. Esta biblioteca utiliza redes neuronales convencionales para predecir la detección de sentimiento en idioma español. El modelo ha sido entrenado con 800.000 valoraciones de usuarios a través de las plataformas de eltenedor, Decathlon, tripadvisor, filmaffinity y ebay, y alcanza un 88% de precisión. 

Su uso no será tan sencillo como los anteriores pasos, pues además de instalar el paquete sentiment-analysis-spanish, se requiere contar con las bibliotecas Keras y Tensorflow en nuestro dispositivo.



from sentiment_analysis_spanish import sentiment_analysis

En este caso, retomaremos el corpus limpio creado en los primeros pasos, y aplicaremos la función de análisis de sentimiento a cada uno de los periódicos, de forma que extraigamos una valoración general a través de la función sentiment() y que nos devuelva una valoración que va desde:
* 0 - más negativo 
* 1 - más positivo

# Recordemos que tenemos por corpus un marco DataFrame de Pandas

corpus_df = pd.read_pickle('corpus.pkl')

#Iniciamos la función se análisis de sentimiento

sentimiento  = sentiment_analysis.SentimentAnalysisSpanish()

'''Definimos qué medio queremos investigar apuntándolo en el DataFrame'''
medio = data.Noticias.loc["Nombre del periódico"]

sentimiento = sentimiento.sentiment(medio)
